+++ assignment1/README.md
@@ -1,4 +1,4 @@
-# wordle-web
+# assignment1
 
 This template should help get you started developing with Vue 3 in Vite.
