+++ assignment1/src/App.vue
@@ -1,3 +1,8 @@
+<script setup lang="ts">
+import { RouterLink, RouterView } from 'vue-router'
+import HelloWorld from './components/HelloWorld.vue'
+</script>
+
 <template>
   <header>
     <img alt="Vue logo" class="logo" src="@/assets/logo.svg" width="125" height="125" />
@@ -7,7 +12,6 @@
 
       <nav>
         <RouterLink to="/">Home</RouterLink>
-        <RouterLink to="/about">Class</RouterLink>
         <RouterLink to="/about">About</RouterLink>
       </nav>
     </div>
@@ -15,11 +19,6 @@
 
   <RouterView />
 </template>
-
-<script setup lang="ts">
-import { RouterLink, RouterView } from 'vue-router'
-import HelloWorld from './components/HelloWorld.vue'
-</script>
 
 <style scoped>
 header {
