+++ assignment1/src/components/icons/IconCommunity.vue
