+++ assignment1/src/components/icons/IconEcosystem.vue
