--- wordle-web/src/views/HomeView.vue
@@ -1,9 +1,0 @@
-<template>
-  <main>
-    <TheWelcome />
-  </main>
-</template>
-
-<script setup lang="ts">
-import TheWelcome from '../components/TheWelcome.vue'
-</script>
